refactor: extract SimpleIme class
