+++ src/options.ts
@@ -1,8 +1,4 @@
 export interface Options {
-  // 允许输入的拼音产长度
+  // TODO: 允许输入的拼音长度
   maxPinyinLength?: number
-  // 是否显示光标
-  cursorMode?: boolean
-  // 是否支持单引号分词
-  singleQuoteDivide?: boolean
 }
